# terraform/openapi.yaml
swagger: "2.0"
info:
  title: "Whales API"
  version: "1.0.0"
basePath: "/api"
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"

paths:
  /health:
    get:
      summary: "Health check endpoint"
      operationId: "health"
      x-google-backend:
        address: "${cloud_run_url}"
        protocol: "h2"
      responses:
        200:
          description: "Service is healthy"

  /data:
    get:
      summary: "Fetch data"
      operationId: "getData"
      x-google-backend:
        address: "${cloud_run_url}"
        protocol: "h2"
      responses:
        200:
          description: "Data retrieved successfully"

  /items:
    get:
      summary: "List all items"
      operationId: "listItems"
      x-google-backend:
        address: "${cloud_run_url}"
        protocol: "h2"
      responses:
        200:
          description: "List of items"
    post:
      summary: "Create new item"
      operationId: "createItem"
      parameters:
        - in: "body"
          name: "item"
          schema:
            type: "object"
            properties:
              title:
                type: "string"
              value:
                type: "number"
      x-google-backend:
        address: "${cloud_run_url}"
        protocol: "h2"
      responses:
        201:
          description: "Item created"

  # Whale Images - Phase 1 Endpoints
  /images/request-upload:
    post:
      summary: "Request presigned URL for image upload"
      operationId: "requestImageUpload"
      parameters:
        - in: "body"
          name: "upload_request"
          required: true
          schema:
            type: "object"
            properties:
              fileName:
                type: "string"
              fileSize:
                type: "integer"
              mimeType:
                type: "string"
              metadata:
                type: "object"
      x-google-backend:
        address: "${cloud_run_url}"
        protocol: "h2"
      responses:
        200:
          description: "Presigned URL generated"
        400:
          description: "Invalid request"

  /images/confirm-upload:
    post:
      summary: "Confirm image upload completion"
      operationId: "confirmImageUpload"
      parameters:
        - in: "body"
          name: "confirm_request"
          required: true
          schema:
            type: "object"
            properties:
              imageId:
                type: "string"
              gcsPath:
                type: "string"
      x-google-backend:
        address: "${cloud_run_url}"
        protocol: "h2"
      responses:
        200:
          description: "Upload confirmed"
        400:
          description: "Invalid request"

  /images/search:
    get:
      summary: "Search whale images by filters"
      operationId: "searchImages"
      parameters:
        - in: "query"
          name: "species"
          type: "string"
        - in: "query"
          name: "region"
          type: "string"
        - in: "query"
          name: "startDate"
          type: "string"
        - in: "query"
          name: "endDate"
          type: "string"
        - in: "query"
          name: "behavior"
          type: "string"
        - in: "query"
          name: "whaleName"
          type: "string"
        - in: "query"
          name: "observerName"
          type: "string"
        - in: "query"
          name: "limit"
          type: "integer"
          default: 20
        - in: "query"
          name: "offset"
          type: "integer"
          default: 0
      x-google-backend:
        address: "${cloud_run_url}"
        protocol: "h2"
      responses:
        200:
          description: "List of whale images"
        400:
          description: "Invalid query parameters"

  /images/{imageId}:
    get:
      summary: "Get image metadata by ID"
      operationId: "getImage"
      parameters:
        - in: "path"
          name: "imageId"
          required: true
          type: "string"
      x-google-backend:
        address: "${cloud_run_url}"
        protocol: "h2"
      responses:
        200:
          description: "Image metadata retrieved"
        404:
          description: "Image not found"
    delete:
      summary: "Delete image by ID"
      operationId: "deleteImage"
      parameters:
        - in: "path"
          name: "imageId"
          required: true
          type: "string"
      x-google-backend:
        address: "${cloud_run_url}"
        protocol: "h2"
      responses:
        204:
          description: "Image deleted"
        404:
          description: "Image not found"